+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 ## DONE
 - V0.0.7
+  - 2022-01-21
+    - [streamlit] add example on cascading dropdown, update streamlit version
   - 2021-12-02
     - [streamlit] add example on preserving state between page changes
   - 2021-11-10